% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blast_formatter}
\alias{blast_formatter}
\title{blast_formatter}
\usage{
blast_formatter(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run blast_formatter
}
\examples{
library(outsider)
blast_formatter <- module_import('blast_formatter',
                                 repo = 'dombennett/om..blast')
blast_formatter('-help')
}
