% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastdb_aliastool}
\alias{blastdb_aliastool}
\title{blastdb_aliastool}
\usage{
blastdb_aliastool(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run blastdb_aliastool
}
\examples{
library(outsider)
blastdb_aliastool <- module_import('blastdb_aliastool',
                              repo = 'dombennett/om..blast')
blastdb_aliastool('-help')
}
