% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastdbcmd}
\alias{blastdbcmd}
\title{blastdbcmd}
\usage{
blastdbcmd(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run blastdbcmd
}
\examples{
library(outsider)
blastdbcmd <- module_import('blastdbcmd', repo = 'dombennett/om..blast..2.7.1')
blastdbcmd('-help')
}
