% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastn}
\alias{blastn}
\title{blastn}
\usage{
blastn(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run blastn
}
\examples{
library(outsider)
blastn <- module_import('blastn', repo = 'dombennett/om..blast')
blastn('-help')
}
