% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastp}
\alias{blastp}
\title{blastp}
\usage{
blastp(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run blastp
}
\examples{
library(outsider)
blastp <- module_import('blastp', repo = 'dombennett/om..blast..2.7.1')
blastp('-help')
}
