% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastx}
\alias{blastx}
\title{blastx}
\usage{
blastx(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run blastx
}
\examples{
library(outsider)
blastx <- module_import('blastx', repo = 'dombennett/om..blast')
blastx('-help')
}
