% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deltablast}
\alias{deltablast}
\title{deltablast}
\usage{
deltablast(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run deltablast
}
\examples{
library(outsider)
deltablast <- module_import('deltablast', repo = 'dombennett/om..blast')
deltablast('-help')
}
