% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeblastdb}
\alias{makeblastdb}
\title{makeblastdb}
\usage{
makeblastdb(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run makeblastdb
}
\examples{
library(outsider)
makeblastdb <- module_import('makeblastdb', repo = 'dombennett/om..blast')
makeblastdb('-help')
}
