% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makembindex}
\alias{makembindex}
\title{makembindex}
\usage{
makembindex(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run makembindex
}
\examples{
library(outsider)
makembindex <- module_import('makembindex',
                             repo = 'dombennett/om..blast')
makembindex('-help')
}
