% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rpstblastn}
\alias{rpstblastn}
\title{rpstblastn}
\usage{
rpstblastn(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run rpstblastn
}
\examples{
library(outsider)
rpstblastn <- module_import('rpstblastn', repo = 'dombennett/om..blast')
rpstblastn('-help')
}
