% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{segmasker}
\alias{segmasker}
\title{segmasker}
\usage{
segmasker(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run segmasker
}
\examples{
library(outsider)
segmasker <- module_import('segmasker', repo = 'dombennett/om..blast')
segmasker('-help')
}
