% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tblastn}
\alias{tblastn}
\title{tblastn}
\usage{
tblastn(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run tblastn
}
\examples{
library(outsider)
tblastn <- module_import('tblastn', repo = 'dombennett/om..blast')
tblastn('-help')
}
