% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{windowmasker}
\alias{windowmasker}
\title{windowmasker}
\usage{
windowmasker(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run windowmasker
}
\examples{
library(outsider)
windowmasker <- module_import('windowmasker',
                              repo = 'dombennett/om..blast..2.7.1')
windowmasker('-help')
}
