% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastdbcmd}
\alias{blastdbcmd}
\title{blastdbcmd}
\usage{
blastdbcmd(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run blastdbcmd
}
\examples{
library(outsider)
blastdbcmd <- module_import('blastdbcmd', repo = 'dombennett/om..blast')
blastdbcmd('-help')
}
