% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blastp}
\alias{blastp}
\title{blastp}
\usage{
blastp(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run blastp
}
\examples{
library(outsider)
blastp <- module_import('blastp', repo = 'dombennett/om..blast')
blastp('-help')
}
