% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert2blastmask}
\alias{convert2blastmask}
\title{convert2blastmask}
\usage{
convert2blastmask(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run convert2blastmask
}
\examples{
library(outsider)
convert2blastmask <- module_import('convert2blastmask',
                                   repo = 'dombennett/om..blast')
convert2blastmask('-help')
}
