% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deltablast}
\alias{deltablast}
\title{deltablast}
\usage{
deltablast(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run deltablast
}
\examples{
library(outsider)
deltablast <- module_import('deltablast', repo = 'dombennett/om..blast..2.7.1')
deltablast('-help')
}
