% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dustmasker}
\alias{dustmasker}
\title{dustmasker}
\usage{
dustmasker(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run dustmasker
}
\examples{
library(outsider)
dustmasker <- module_import('dustmasker', repo = 'dombennett/om..blast')
dustmasker('-help')
}
