% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeprofiledb}
\alias{makeprofiledb}
\title{makeprofiledb}
\usage{
makeprofiledb(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run makeprofiledb
}
\examples{
library(outsider)
makeprofiledb <- module_import('makeprofiledb',
                               repo = 'dombennett/om..blast')
makeprofiledb('-help')
}
