% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{psiblast}
\alias{psiblast}
\title{psiblast}
\usage{
psiblast(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run psiblast
}
\examples{
library(outsider)
psiblast <- module_import('psiblast', repo = 'dombennett/om..blast')
psiblast('-help')
}
