% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rpsblast}
\alias{rpsblast}
\title{rpsblast}
\usage{
rpsblast(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run rpsblast
}
\examples{
library(outsider)
rpsblast <- module_import('rpsblast', repo = 'dombennett/om..blast..2.7.1')
rpsblast('-help')
}
