% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tblastx}
\alias{tblastx}
\title{tblastx}
\usage{
tblastx(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run tblastx
}
\examples{
library(outsider)
tblastx <- module_import('tblastx', repo = 'dombennett/om..blast..2.7.1')
tblastx('-help')
}
