% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{windowmasker}
\alias{windowmasker}
\title{windowmasker}
\usage{
windowmasker(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments}
}
\description{
Run windowmasker
}
\examples{
library(outsider)
windowmasker <- module_import('windowmasker',
                              repo = 'dombennett/om..blast')
windowmasker('-help')
}
